+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This CHANGELOG details important changes made in each version of the
 - The Python SDK generated for a provider now supports calling `SomeResource.get(...)` to create a
   resource with the state of existing cloud resource.
 - Generate named nested types with doc comments instead of anonymous inline expansions for TypeScript.
+- Set `DeleteBeforeReplace` for resources that are not auto-named.
 
 ## v0.18.3 (Released June 20, 2019)